+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 
 - ... <!-- Add new lines here. -->
 
+## 0.1.1 — 0.1.3
+
+_2022-02-20_
+
+- feat: Add `color` helpers
+- feat: Add angle unit function `deg()` — and converters from `turn`, `rad`
+  and `grad`
+- feat: Add new `VariableOptions` methods: `resolveType`, `isColor`
+- feat: Add `VariablePrinter.type`
+- feat: Log warnings on wonky `scoped()` prefixes (in dev mode)
+- docs: Fix error in code example
+
 ## 0.1.0
 
 _2022-02-17_